+++ test/class.async.test.tsx
@@ -18,7 +18,7 @@ class ResolvedState extends React.Component {
         <button data-testid="async">{this.statex.dynamic.value}</button>
         <button data-testid="async-valueof">{this.statex.dynamic+''}</button>
       </div>
-    )
+    );
   }
 }
 
@@ -35,8 +35,8 @@ class RejectedState extends React.Component {
       <div>
         <button data-testid="async">{this.statex.dynamic.value}</button>
       </div>
-    )
-  };
+    );
+  }
 }
 
 class DefaultState extends React.Component {
@@ -52,14 +52,14 @@ class DefaultState extends React.Component {
       <div>
         <button data-testid="default">{this.statex.dynamic.value}</button>
       </div>
-    )
+    );
   }
 }
 
 it('async: should re-render when dynamic state resolved', async () => {
   const { getByTestId } = render(<ResolvedState />);
 
-  const btn = getByTestId("async");
+  const btn = getByTestId('async');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('done'));
@@ -68,7 +68,7 @@ it('async: should re-render when dynamic state resolved', async () => {
 it('async: should re-render when dynamic state rejected', async () => {
   const { getByTestId } = render(<RejectedState />);
 
-  const btn = getByTestId("async");
+  const btn = getByTestId('async');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('failed'));
@@ -77,7 +77,7 @@ it('async: should re-render when dynamic state rejected', async () => {
 it('async: should get the same state value by valueOf', async () => {
   const { getByTestId } = render(<ResolvedState />);
 
-  const btn = getByTestId("async-valueof");
+  const btn = getByTestId('async-valueof');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('done'));
@@ -86,7 +86,7 @@ it('async: should get the same state value by valueOf', async () => {
 it('async: should get a default value if not specified initialValue or fallbackValue', async () => {
   const { getByTestId } = render(<DefaultState />);
 
-  const btn = getByTestId("default");
+  const btn = getByTestId('default');
   expect(btn.innerHTML).toBe('');
 
   await waitFor(() => expect(btn.innerHTML).toBe(''));