+++ test/function.array.test.tsx
@@ -27,13 +27,13 @@ function State() {
         {state.array.join(',')}
       </button>
     </div>
-  )
+  );
 }
 
 it('array-push: should re-render when array pushed', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-push");
+  const btn = getByTestId('array-push');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);
@@ -46,7 +46,7 @@ it('array-push: should re-render when array pushed', async () => {
 it('array-pop: should re-render when array poped', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-pop");
+  const btn = getByTestId('array-pop');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);
@@ -59,7 +59,7 @@ it('array-pop: should re-render when array poped', async () => {
 it('array-shift: should re-render when array shifted', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-shift");
+  const btn = getByTestId('array-shift');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);
@@ -72,7 +72,7 @@ it('array-shift: should re-render when array shifted', async () => {
 it('array-unshift: should re-render when array unshifted', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-unshift");
+  const btn = getByTestId('array-unshift');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);
@@ -85,7 +85,7 @@ it('array-unshift: should re-render when array unshifted', async () => {
 it('array-splice: should re-render when array spliced', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-splice");
+  const btn = getByTestId('array-splice');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);
@@ -98,7 +98,7 @@ it('array-splice: should re-render when array spliced', async () => {
 it('array-assign: should re-render when array assigned', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("array-assign");
+  const btn = getByTestId('array-assign');
   expect(btn.innerHTML).toBe('0,1,2,3');
 
   fireEvent.click(btn);