+++ test/function.async.test.tsx
@@ -16,7 +16,7 @@ function ResolvedState() {
       <button data-testid="async">{state.dynamic.value}</button>
       <button data-testid="async-valueof">{state.dynamic+''}</button>
     </div>
-  )
+  );
 }
 
 function RejectedState() {
@@ -30,7 +30,7 @@ function RejectedState() {
     <div>
       <button data-testid="async">{state.dynamic.value}</button>
     </div>
-  )
+  );
 }
 
 function DefaultState() {
@@ -44,13 +44,13 @@ function DefaultState() {
     <div>
       <button data-testid="default">{state.dynamic.value}</button>
     </div>
-  )
+  );
 }
 
 it('async: should re-render when dynamic state resolved', async () => {
   const { getByTestId } = render(<ResolvedState />);
 
-  const btn = getByTestId("async");
+  const btn = getByTestId('async');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('done'));
@@ -59,7 +59,7 @@ it('async: should re-render when dynamic state resolved', async () => {
 it('async: should re-render when dynamic state rejected', async () => {
   const { getByTestId } = render(<RejectedState />);
 
-  const btn = getByTestId("async");
+  const btn = getByTestId('async');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('failed'));
@@ -68,7 +68,7 @@ it('async: should re-render when dynamic state rejected', async () => {
 it('async: should get the same state value by valueOf', async () => {
   const { getByTestId } = render(<ResolvedState />);
 
-  const btn = getByTestId("async-valueof");
+  const btn = getByTestId('async-valueof');
   expect(btn.innerHTML).toBe('loading...');
 
   await waitFor(() => expect(btn.innerHTML).toBe('done'));
@@ -77,7 +77,7 @@ it('async: should get the same state value by valueOf', async () => {
 it('async: should get a default value if not specified initialValue or fallbackValue', async () => {
   const { getByTestId } = render(<DefaultState />);
 
-  const btn = getByTestId("default");
+  const btn = getByTestId('default');
   expect(btn.innerHTML).toBe('');
 
   await waitFor(() => expect(btn.innerHTML).toBe(''));