+++ test/function.exception.test.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { render } from '@testing-library/react';
 import React from 'react';
 import { stateProxy } from '../src/index';
@@ -7,21 +8,21 @@ it('number: should throw an exception when stateTarget was not an object', async
   const original = console.error;
   console.error = jest.fn();
   try {
-    render(<State1 />)
+    render(<State1 />);
   } catch(e) {
     expect(e.message).toContain('The [stateTarget] must be an object.');
   }
 
   const State2 = () => stateProxy('str' as any);
   try {
-    render(<State2 />)
+    render(<State2 />);
   } catch(e) {
     expect(e.message).toContain('The [stateTarget] must be an object.');
   }
 
   const State3 = () => stateProxy(false as any);
   try {
-    render(<State3 />)
+    render(<State3 />);
   } catch(e) {
     expect(e.message).toContain('The [stateTarget] must be an object.');
   }