+++ test/function.number.test.tsx
@@ -12,13 +12,13 @@ function State() {
     <div>
       <button data-testid="number" onClick={() => state.counter++}>{state.counter}</button>
     </div>
-  )
+  );
 }
 
 it('number: should re-render when counter chagned', async () => {
   const { getByTestId } = render(<State />);
 
-  const btn = getByTestId("number");
+  const btn = getByTestId('number');
   expect(btn.innerHTML).toBe('0');
 
   fireEvent.click(btn);