+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Changelog
+[2021-12-29]
+1. subscribed re-render
+
 [2021-11-18]
 1. add `stateWrapper` API.